+++ auto_kappa/plot/bandos.py
@@ -79,16 +79,7 @@ def plot_bandos_for_different_sizes(
     msg = " Output %s" % figname
     logger.info(msg)
     
-    ### old version
-    # plot_bandos(
-    #         directory=almcalc1.out_dirs["harm"]["bandos"],
-    #         prefix=almcalc1.prefix,
-    #         directory2=almcalc2.out_dirs["harm"]["bandos"],
-    #         prefix2=almcalc2.prefix,
-    #         fig_labels=[lab1, lab2],
-    #         figname=figname.replace(os.getcwd(), "."),
-    #         )
-
+    
 def set_xticks_labels(ax, kmax, ksym, labels):
     """ """    
     dk_all = kmax