+++ examples/ak_relax/run.sh
@@ -1,0 +1,5 @@
+python make_vasp_files.py \
+    --poscar ./Structure/POSCAR.Si \
+    --outdir ./out \
+    --config vasp_config.yaml
+